+++ plugin/config/config.php
@@ -4,4 +4,7 @@
 
 return [
     'version' => '0.0.1',
+    'enabled_crons.hMarfaImport' => true,
+    'hmarfa.import.script_file' => '/hMarfaImport.sql',
+    'hmarfa.import.errors' => PLUGIN_ISERVICE_CACHE_DIR . "/hMarfaImportErrors",
 ];