+++ plugin/inc/utils/toolbox.class.php
@@ -56,29 +56,6 @@ class ToolBox
         };
     }
 
-    public static function getQueryResult($query, $id_field = 'id', ?\DBmysql $db = null): bool|array
-    {
-        if ($db === null) {
-            global $DB;
-            $db = $DB;
-        }
-
-        $query_result = [];
-        if (false === ($result = $db->query($query)) || $result === true || !$db->numrows($result)) {
-            return is_bool($result) ? $result : $query_result;
-        }
-
-        while ($data = $db->fetchAssoc($result)) {
-            if (isset($data[$id_field])) {
-                $query_result[$data[$id_field]] = $data;
-            } else {
-                $query_result[] = $data;
-            }
-        }
-
-        return $query_result;
-    }
-
     public static function getHtmlSanitizedValue($value): string
     {
         return preg_replace('((?![\w\-]).)', '-', trim($value));