+++ plugin/install/sql/delete_tables.sql
@@ -1,3 +1,21 @@
-DROP TABLE IF EXISTS `glpi_plugin_iservice_configs`;
+drop table if exists `glpi_plugin_iservice_configs`;
 
-DROP TABLE IF EXISTS `glpi_plugin_iservice_importmappings`;
+drop table if exists `glpi_plugin_iservice_importmappings`;
+
+drop table if exists `hmarfa_facrind`;
+
+drop table if exists `hmarfa_facturi`;
+
+drop table if exists `hmarfa_firme`;
+
+drop table if exists `hmarfa_gestiuni`;
+
+drop table if exists `hmarfa_incpla`;
+
+drop table if exists `hmarfa_lotm`;
+
+drop table if exists `hmarfa_nommarfa`;
+
+drop table if exists `hmarfa_tran`;
+
+drop view if exists hmarfa_total_facturi;